move groups methods to their own files
